Refactor: 상세 페이지 로그인 경고 모달 위치 수정
