+++ src/app/_component/common/layout/Dialog.tsx
@@ -36,7 +36,7 @@ const Dialog = ({
       <div className="z-[500] fixed top-0 left-0 w-full h-full  bg-black/60 cursor-pointer" />
       {type === "confirm" && (
         <dialog
-          className="z-[550] w-[87.2%] web:w-[327px] h-[176px] mx-auto absolute top-[30%] rounded-md bg-white flex flex-col justify-evenly shadow-md "
+          className="z-[550] w-[87.2%] web:w-[327px] h-[176px] mx-auto fixed top-[30%] rounded-md bg-white flex flex-col justify-evenly shadow-md "
           ref={ref}
         >
           {theme === "withdraw" && (
